fix(stats): use week data for streak pills
Fixes https://github.com/trakt/trakt-web/issues/2236
