+++ projects/client/src/lib/sections/stats/_internal/models/HeatmapIntensity.ts
@@ -1,0 +1,1 @@
+export type HeatmapIntensity = 0 | 1 | 2 | 3 | 4;
